Draft.

+++ shortcuts/factory/factoryCreate_here.sh
@@ -2,14 +2,27 @@
 _here_dockerfile-ubiquitous() {
 
 # DANGER: ONLY in Docker container in CI environment !
-local currentDirectory=$(realpath --relative-to="$PWD" "$scriptAbsoluteFolder")
-[[ "$CI" != "" ]] && cat << CZXWXcRMTo8EmM8i4d
+if [[ "$CI" != "" ]]
+then
+
+mkdir -p "$safeTmp"/repo/ubiquitous_bash
+cp -a "$scriptAbsoluteFolder"/.git "$safeTmp"/repo/ubiquitous_bash/
+( cd "$safeTmp"/repo/ubiquitous_bash ; "$scriptAbsoluteLocation"/ubiquitous_bash.sh _gitBest reset --hard ; git submodule update --init --recursive )
+
+#local currentDirectoy=$(realpath --relative-to="$PWD" "$scriptAbsoluteFolder")
+local currentDirectoy=$(realpath --relative-to="$PWD" "$safeTmp"/repo/ubiquitous_bash)
+
+
+cat << CZXWXcRMTo8EmM8i4d
 
 RUN rm -rf /workspace/ubiquitous_bash
 RUN mkdir -p /workspace
 COPY $currentDirectory /workspace/ubiquitous_bash
 
 CZXWXcRMTo8EmM8i4d
+
+export safeToDeleteGit="true"
+fi
 
 cat << 'CZXWXcRMTo8EmM8i4d'
