+++ ubiquitous_bash.sh
@@ -36,7 +36,7 @@ _ub_cksum_special_derivativeScripts_contents() {
 #export ub_setScriptChecksum_disable='true'
 ( [[ -e "$0".nck ]] || [[ "${BASH_SOURCE[0]}" != "${0}" ]] || [[ "$1" == '--profile' ]] || [[ "$1" == '--script' ]] || [[ "$1" == '--call' ]] || [[ "$1" == '--return' ]] || [[ "$1" == '--devenv' ]] || [[ "$1" == '--shell' ]] || [[ "$1" == '--bypass' ]] || [[ "$1" == '--parent' ]] || [[ "$1" == '--embed' ]] || [[ "$1" == '--compressed' ]] || [[ "$0" == "/bin/bash" ]] || [[ "$0" == "-bash" ]] || [[ "$0" == "/usr/bin/bash" ]] || [[ "$0" == "bash" ]] ) && export ub_setScriptChecksum_disable='true'
 export ub_setScriptChecksum_header='2591634041'
-export ub_setScriptChecksum_contents='413146563'
+export ub_setScriptChecksum_contents='1344665118'
 
 # CAUTION: Symlinks may cause problems. Disable this test for such cases if necessary.
 # WARNING: Performance may be crucial here.
@@ -26948,14 +26948,27 @@ alias l=_l
 _here_dockerfile-ubiquitous() {
 
 # DANGER: ONLY in Docker container in CI environment !
-local currentDirectory=$(realpath --relative-to="$PWD" "$scriptAbsoluteFolder")
-[[ "$CI" != "" ]] && cat << CZXWXcRMTo8EmM8i4d
+if [[ "$CI" != "" ]]
+then
+
+mkdir -p "$safeTmp"/repo/ubiquitous_bash
+cp -a "$scriptAbsoluteFolder"/.git "$safeTmp"/repo/ubiquitous_bash/
+( cd "$safeTmp"/repo/ubiquitous_bash ; "$scriptAbsoluteLocation"/ubiquitous_bash.sh _gitBest reset --hard ; git submodule update --init --recursive )
+
+#local currentDirectoy=$(realpath --relative-to="$PWD" "$scriptAbsoluteFolder")
+local currentDirectoy=$(realpath --relative-to="$PWD" "$safeTmp"/repo/ubiquitous_bash)
+
+
+cat << CZXWXcRMTo8EmM8i4d
 
 RUN rm -rf /workspace/ubiquitous_bash
 RUN mkdir -p /workspace
 COPY $currentDirectory /workspace/ubiquitous_bash
 
 CZXWXcRMTo8EmM8i4d
+
+export safeToDeleteGit="true"
+fi
 
 cat << 'CZXWXcRMTo8EmM8i4d'
 
@@ -27775,6 +27788,9 @@ __factoryCreate_sequence_runpod-heavy() {
     docker tag runpod-heavy "$DOCKER_USER"/runpod-heavy:latest
 
     #docker push user/runpod-heavy:latest
+
+    #export safeToDeleteGit="true"
+    _safeRMR "$safeTmp"/repo
 
     _stop
 }
